refactor: move remaining artifactory registry tests to generic test class
