+++ dev/fattest.simplicity/src/componenttest/containers/registry/InternalRegistry.java
@@ -25,9 +25,9 @@ public class InternalRegistry extends Registry {
      * Expect this to be set on remote build machines. Local build machines will
      * likely not be able to use this registry.
      */
-    public static final String REGISTRY = "docker_registry.server";
-    public static final String REGISTRY_USER = "docker_registry.user";
-    public static final String REGISTRY_PASSWORD = "docker_registry.password";
+    private static final String REGISTRY = "docker_registry.server";
+    private static final String REGISTRY_USER = "docker_registry.user";
+    private static final String REGISTRY_PASSWORD = "docker_registry.password";
 
     private static final String DEFAULT_REGISTRY = ""; //Blank registry is the default setting
